latest from coda-oss

+++ externals/coda-oss/modules/c++/numpyutils/unittests/test_num_elements.cpp
@@ -37,7 +37,7 @@ TEST_CASE(testGetNumElements)
     // so we don't have anyone to do the cleanup work.
     std::vector<int> data(50);
     PyObject* pyArrayObject = numpyutils::toNumpyArray(5, 10, NPY_INT, data.data());
-    TEST_ASSERT_EQ(50, numpyutils::getNumElements(pyArrayObject));
+    TEST_ASSERT_EQ(static_cast<size_t>(50), numpyutils::getNumElements(pyArrayObject));
 
     PyObject* nonArrayObject = PyLong_FromLong(123L);
     TEST_THROWS(numpyutils::getNumElements(nonArrayObject));
