+++ externals/coda-oss/modules/c++/sys/unittests/test_path.cpp
@@ -58,7 +58,7 @@ TEST_CASE(testPathMerge)
     std::vector<std::string> paths;
     const auto splitResult = os.splitEnv("PATH", paths);
     TEST_ASSERT_TRUE(splitResult);
-    TEST_ASSERT_GREATER(paths.size(), 0);
+    TEST_ASSERT_GREATER(paths.size(), static_cast<size_t>(0));
 
     auto path = find_directory(paths);
     TEST_ASSERT_TRUE(fs::is_directory(path));
@@ -70,7 +70,7 @@ TEST_CASE(testPathMerge)
 
     bool isAbsolute;
     auto components = sys::Path::separate(path, isAbsolute);
-    TEST_ASSERT_GREATER(components.size(), 0);
+    TEST_ASSERT_GREATER(components.size(), static_cast<size_t>(0));
     auto result = sys::Path::merge(components, isAbsolute);
     TEST_ASSERT_EQ(result, path);
     TEST_ASSERT_TRUE(fs::is_directory(result));
